Add LevenshtrieSet instead of LevenshtrieMultiMap
LevenshtrieMultiMap was a broken abstraction - it was unclear whether the multiple values had set semantics (no duplicates) or list semantics (duplicates allowed). This meant that the API was not really helpful and that performance was very sluggish where set semantics was necessary.

Thus LevenshtrieSet was born. This new data structure organises results like a HashSet<T> does by using buckets. this drastically improves indexing performance where words may map to values and deduplication is necessary, which is a common usage of tries.

Some refactoring was done to enable a reintroduction of LevenshtrieList if we ever choose to do so.